Initial Commit of experimental module
For anonymous users, redirects from desired node to parent item (via path) if access is denied.
The alias for the requested path is obtained, and the last portion is chopped off to obtain the parent path.
Configuration for exclusions to this is added to site info config. Exclusions are redirect to home.